tom: resize the disk containing nix storage after an unmounting
youch this was a painful mistake. do not unmount the /nix/store or core
commands are lost!!

i was hoping to increase the disk size for the /nix/store so mounted
a disk, copied the store, updated the configuration.nix, and rebuilt.

after unmounting this disk things went bad. no sudo to edit configs.

rebooting failed for that generation so with the layout of disks and
uuids being different, reinstalling was found to be the quickest fix.

no data was lost and new commands were learned.